Fix empty item creation issue
Currently there is an issue which allows kind of creation of empty
items. When the focus is blurred on input of a new item with empty text,
the item will remain in the UI even though it isn't sent to the API.
This is because of an test where the update/creation event is not
dispatched when the text hasn't been changed. By tweaking this test a
bit, and getting rid of empty items in `create` event listener, this
annoying bug can be fixed.